requirements.txt
pyyaml>=6.0.1
rich>=13.7.0
python-dateutil>=2.8.2
yt-dlp>=2023.12.30
openai>=1.3.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
deepgram-sdk>=2.12.0 